requirements.txt
requests
pandas
tqdm
python-dotenv
PyGithub
matplotlib
seaborn
scipy